+++ addons/im_livechat/static/src/legacy/widgets/public_livechat_view/public_livechat_view.js
